# benchmark: also report the inverse of the performance (ops/s)