Report the performance in operations per second too, like so:

```
---------------------------
  Kyber512 | (1000 calls)
---------------------------
Keygen: 5.569 s (179.571/s)
Enc: 8.278 s (120.799/s)
Dec: 8.278 s (74.929/s)
---------------------------
  Kyber768 | (1000 calls)
---------------------------
Keygen: 8.761 s (114.138/s)
Enc: 12.793 s (78.17/s)
Dec: 12.793 s (49.208/s)
---------------------------
  Kyber1024 | (1000 calls)
---------------------------
Keygen: 13.473 s (74.22/s)
Enc: 18.706 s (53.46/s)
Dec: 18.706 s (34.171/s)
```