@GiacomoPope 2024-07-18
While we're changing this, maybe we could use something like `tabulate` (or just do our own string formatting) to make this table prettier?

```
     Kyber512 - (1000 iterations)
-----------------------------------
Algorithm     | time (s) | ops / s
-----------------------------------
Keygen        | 5.569    | 180
Encapsulation | 8.278 s  | 121
Decapsulation | 8.278 s  | 75
```

Probably worth rounding to an integer for the ops / s, do you agree?