@tomato42 2024-07-18
that's literally the second thing I wanted to propose :grin: 

but I was thinking of doing a transpose of it:
```
Group  |  keygen  | keygen/s | encapsulation | encapsulation/s | decapsulation | decapsulation/s 
---------------------------------------------------------------------------------------------------------------------------------------------
Kyber512 | 123 ...
```

(i.e. something that `openssl speed` outputs)