@GiacomoPope 2024-07-18
Yeah this seems good. The current output was something I wrote in 3 minutes to just check what's happening! I also thing average time is one metric, but minimum / maximum times is maybe interesting too.