@GiacomoPope 2024-07-18
haha yeah, hence the conservative README:

```
TODO: Better benchmarks? Although this was never about speed haha.

For now, here are some approximate benchmarks:
```