@GiacomoPope 2024-07-18
I would swap `group` with `parameters` or `param set` just because it's not a maths group and it feels weird to call it a group.

I would also report time in ms rather than seconds so something like:

```
-------------------------------------------------------------------
 group     | keygen | keygen/s | encap | encap/s | decap | decap/s 
-------------------------------------------------------------------
 Kyber512  |  5.84ms   171.21    8.67ms  115.28   13.96ms     71.65
 Kyber768  |  8.39ms   119.18   12.22ms   81.86   19.38ms     51.60
 Kyber1024 | 12.75ms    78.44   17.74ms   56.37   27.74ms     36.05
 ```