@tomato42 2024-07-18
I've used "group" as that's the name in use in TLS: "key exchange group", to encompass both the FFDH prime, ECDH curve, and now PQC parameters. Changed to params