@GiacomoPope 2024-07-18
Cool. Looks good to me. I'll update the benchmark in the README tonight